Improve body parameter for openapi_3 to swagger_2 conversion (#205)
* openapi_3 to swagger_2: Fix invalid body parameter

An OAS v3 spec with an operation to send a file in a request body, as
described in [Considerations for File Uploads][1]:

    requestBody:
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary

api-spec-converter converts this to:

    consumes:
      - application/octet-stream
    parameters:
      - in: formData
        name: body
        type: file
    responses:
      '204':
        description: Success
    operationId: uploadFile

Which is invalid because ["if type is `file`, the consumes MUST be either
`multipart/form-data`, `application/x-www-form-urlencoded` or both"][2].

Instead, preserve the schema from the OAS3 spec, if present, or use
`type: "string"` with `format: "binary"` as recommended in
OAI/OpenAPI-Specification#326.  Also, don't copy the non-schema
properties, which are invalid for `in: "body"`.

[1]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#considerations-for-file-uploads
[2]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#fixed-fields-7

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* openapi_3 to swagger_2: Support more+multiple types

The transformation for application/octet-stream should work for any
media type.  Instead of failing with 'unsupported request body media
type', apply the transformation.

Additionally, include all media types (but not media ranges, which are
not allowed in Swagger 2) from `content` in `consumes`.  This avoids
losing information for operations which consume multiple types (e.g.
`image/gif` and `image/png`).

If `consumes` would be empty because the OAS3 content object only
contained media ranges, use `application/octet-stream` as a generic
catch-all type.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* openapi_3 to swagger_2: Preserve all request body types

As with upload types, operations may consume both formData types or
multiple JSON types.  Preserve these as well.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>