openapi_3 to 2: Convert multiple response types (#207)
In OAS v3, the content property of the Response Object maps any number
of media ranges to Media Type Objects.  The previous implementation
would only copy `application/json` (with fallback to the first media
range).  This would lose information about operations which can return
multiple types and produce an invalid specification for content starting
with a media range (since ranges are not allowed in Swagger 2).

Instead, include all types from `content` in `produces` and `examples`,
mapping ranges to `application/octet-stream` as a generic catch-all
type.  Since Swagger 2 only supports a single schema, the first schema
for a JSON type is used, with fallback to the first of any type.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>